pyproject.toml
[tool.poetry]
name = "binance-mauricio"
version = "0.1.0"
description = "herramienta para administrar 1 o mas cuentas en binance"
authors = ["EK"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.1"
typer = "^0.4.0"
python-binance = "^1.0.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
